Fix initialisation of map report layer
Following the fix to OpenLayers in the media repository,
https://github.com/Indicia-Team/media/commit/730d9b5800c1e2c5376fec3517266799cd472794,
the behaviour of OpenLayers changed. If the vector layer was drawn
then zoomed to extents the first drawn features were not erased.

This could be worked around by waiting for the 'tilesloaded' event but
is solved here by removing the layer until after the call to
zoomToExtent.