Simplified check for tab alias
TabAlias can be passed to get_control_methods where controls are being
loaded onto a tab. This change passes the alias through if available,
rather than only when the interface style is not one page (which was
checked for anyway). This allows the use of tab alias to be a bit more
fine controlled.