Replaced $node with $nid
For Drupal version compatibility, client code should not be dependent
on node object structure. Therefore we pass $nid around instead as
generally we only need the node ID.
This change already implemented for most code, this commit is to finish
the task for the built in grids in dynamic forms.