Fixes async issue when deleting count values
If a count value existed for a section, then was quickly deleted and
re-input, the success handler for the deletion would not realise the
post was a deletion because the form value has already changed back.
This resulted in the IDs of a deleted record being stored in the hidden
inputs which then get used for future posts, resulting in the warehouse
complaining as it can’t process edits to a deleted record.
Now fixed.