Fixes problem if multiple dynamic report explorers on one page
Previously, if you used something like Drupal panels to add multiple
dynamic report explorers to a single page, then the $reportCount static
variable kept track of the count of reports on the entire page. This
was then used to get the correct set of column configurations from the
indexed array of column configurations for each dynamic report
explorer, but the assumption was made that the reportCount started at
zero which was not always true.