Remove send_file_to_warehouse protection
DP7 Indicia API module needs to send the photos to the warehouse and since it does not use data_entry_helper::extract_media_data it needs to be able to upload the images to the warehouse. I can just copy the function code and use it in the new module, which works fine but I don't think I want to duplicate the code.