Fixes a problem with group input forms
If a form had a taxon filter set, but was loaded for a group then the
group’s taxonomic filter can overwrite the form’s taxonomic filter.
This should not happen when the filter doesn’t have a taxon filter, but
it was due to a mistake in the logic.