Fixes #49 - report map doesn't respect template
Contains a relatively small amount of code to implement the enhancement.
However, includes many style changes and comments to improve legibility.
Most significant alteration is to use of the $colsToInclude array which
was previously empty unless specifically used to constrain columns sent
to JavaScript. Now it is always employed to list output columns. This is
because the code also includes a change to prevent output of columns
that are marked as not visible (unless required for mapping, e.g. geom)