EBMS Issue 26
When saving a site initially, the total transect length is not filled
in, so the first time it is updated when saving a section, it creates
the attribute value. The name of the field must be updated to hold the
new attribute value id, otherwise if it is updated again without
refreshing the page another attribute value record will be created,
rather than updating the exsting one.