Add a CQL filter option
This prebuilt form is configured with a report to list species
which is displayed beside a map showing records of a selected
species. For consistent behaviour, the map and the report need to
be filtered in a similar way. The CQL to filter the map was hard
coded. This adds an option allowing it to be customised.