Template replacements use blank string when object or array passed, instead of skipping.
In the old code, if you had a multi-value attribute then the default would be []. If
this is used against a standard text input, the {default} replacement token never
gets replaced as the code skips objects and arrays. So {default} appears in the control.
Now objects and arrays are replaced in the tokens using an empty string instead.