Fix case where predefined mappings for columns with non alpha characters fails
See https://github.com/BiologicalRecordsCentre/iRecord/issues/717.
Column key passed through in this case was not calculated in the same way
as the rest of the code, so non-alpha characters were being left in place
inconsistently. Then, when the mappings are cleaned up later, this column
gets discarded and the columns get out of sync.