Fix problems with str_replace on templates
Because of left to right operation of str_replace on array searches,
database values containing things that might be interpreted as
tokens such as {id} were getting replaced by the control ID. Fixes that
and a problem with autocompletes introduced by the fix.
An example was setting Redirect to page after successful data entry
to contain a URL with {id} in it that should be replaced with the ID
query param. Editing an existing form with this config was breaking on
save.